fix: recover `convert` proof in Geometry.Euclidean.Inversion (#4421)
During porting a `convert` proof was converted to `calc`. This switches to a more direct translation of the original.

We need `using` now because `convert` is happy to descend into the expressions and equate `(HMul.hMul : ℝ → ℝ → ℝ) = (HDiv.hDiv : ℝ → ℝ → ℝ)` since these involve the same types. The old `convert` wouldn't do this because it used simp's congr lemmas, and these require that the functions be defeq rather than just equal.